Divide BuildEventsExtension to share utils methods
