+++ src/main/java/com/lassekoskela/maven/BuildEventsExtension.java
@@ -1,31 +1,22 @@
 package com.lassekoskela.maven;
 
 import java.io.File;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 import org.apache.maven.AbstractMavenLifecycleParticipant;
 import org.apache.maven.MavenExecutionException;
-import org.apache.maven.execution.ExecutionListener;
-import org.apache.maven.execution.MavenExecutionRequest;
 import org.apache.maven.execution.MavenSession;
-import org.apache.maven.model.Profile;
 import org.codehaus.plexus.component.annotations.Component;
 import org.codehaus.plexus.component.annotations.Requirement;
 import org.codehaus.plexus.logging.Logger;
 
 import com.lassekoskela.maven.buildevents.BuildEventListener;
 import com.lassekoskela.maven.buildevents.BuildEventLog;
-import com.lassekoskela.maven.buildevents.ExecutionListenerChain;
 import com.lassekoskela.maven.logging.ConsoleLog;
 import com.lassekoskela.maven.logging.FileLog;
 import com.lassekoskela.maven.logging.Log;
 
-import static org.apache.commons.lang3.StringUtils.*;
-
 @Component(role = AbstractMavenLifecycleParticipant.class, hint = "buildevents")
-public class BuildEventsExtension extends AbstractMavenLifecycleParticipant {
+public class BuildEventsExtension extends MavenExtension {
 
 	static final String OUTPUT_MODE = "duration.output";
 	static final String OUTPUT_FILE = "duration.output.file";
@@ -37,55 +28,18 @@ public class BuildEventsExtension extends AbstractMavenLifecycleParticipant {
 	Logger logger;
 
 	@Override
-	public void afterProjectsRead(MavenSession session)
-			throws MavenExecutionException {
+	public void afterProjectsRead(MavenSession session) throws MavenExecutionException {
 		Log log = resolveLogDestination(session);
 
-        if (!isActivationProfilesPropertySet(session))
+        if (!isActivationProfilesPropertySet(session, ACTIVATION_PROFILE_KEY))
             log.info("Use " + ACTIVATION_PROFILE_KEY + " property to set in which profiles to run maven-build-utils.");
 
-        if (shouldBeActive(session)) {
+        if (shouldBeActive(session, ACTIVATION_PROPERTY_KEY, ACTIVATION_PROFILE_KEY)) {
             BuildEventListener listener = createListener(log);
             registerExecutionListener(session, listener);
         }
     }
 
-    private boolean shouldBeActive(MavenSession session) {
-        boolean shouldBeActive = Boolean.valueOf(getActivationProperty(session));
-        if (shouldBeActive) {
-        	return shouldBeActive;
-        }
-
-        List<String> activationProfiles = listActivationProfiles(session);
-        for (String currentlyActive : getAllActiveProfileNames(session)) if (activationProfiles.contains(currentlyActive)) shouldBeActive = true;
-        return shouldBeActive;
-    }
-
-    protected List<String> listActivationProfiles(MavenSession session) {
-        return Arrays.asList(split(deleteWhitespace(getActivationProfilesProperty(session)), ','));
-    }
-    
-    protected static String getActivationProfilesProperty(MavenSession session) {
-        return session.getCurrentProject().getProperties().getProperty(ACTIVATION_PROFILE_KEY, "default");
-    }
-
-    
-    private String getActivationProperty(MavenSession session) {
-        return session.getCurrentProject().getProperties().getProperty(ACTIVATION_PROPERTY_KEY, Boolean.FALSE.toString());
-    }
-
-    protected static boolean isActivationProfilesPropertySet(MavenSession session) {
-        return !isBlank(session.getCurrentProject().getProperties().getProperty(ACTIVATION_PROFILE_KEY));
-    }
-
-    protected List<String> getAllActiveProfileNames(MavenSession session) {
-        List<String> names = new ArrayList<String>();
-        for (Profile profile : session.getCurrentProject().getActiveProfiles()) {
-            names.add(profile.getId());
-        }
-        return names;
-    }
-    
 	protected BuildEventListener createListener(Log log) {
 		return new BuildEventListener(new BuildEventLog(log));
 	}
@@ -110,17 +64,5 @@ public class BuildEventsExtension extends AbstractMavenLifecycleParticipant {
 		}
 		String buildDir = session.getExecutionRootDirectory();
 		return new FileLog(new File(buildDir, file));
-	}
-
-	private String getProperty(MavenSession s, String key, String defaultValue) {
-		return s.getUserProperties().getProperty(key, defaultValue);
-	}
-
-	protected void registerExecutionListener(MavenSession session,
-			ExecutionListener listener) {
-		MavenExecutionRequest request = session.getRequest();
-		ExecutionListener original = request.getExecutionListener();
-		ExecutionListener chain = new ExecutionListenerChain(original, listener);
-		request.setExecutionListener(chain);
 	}
 }
